chore: linux getDiskSize()
