+++ core/src/main/c/share/files.c
@@ -24,7 +24,6 @@
 
 #include <unistd.h>
 #include <sys/stat.h>
-#include <sys/statvfs.h>
 #include <fcntl.h>
 #include <sys/file.h>
 #include <sys/mman.h>
@@ -33,10 +32,6 @@
 #ifdef __APPLE__
 
 #include <sys/time.h>
-
-#else
-
-#include <utime.h>
 
 #endif
 
@@ -252,17 +247,6 @@ JNIEXPORT jboolean JNICALL Java_io_questdb_std_Files_setLastModified
     return (jboolean) (utimes((const char *) lpszName, t) == 0);
 }
 
-#else
-
-JNIEXPORT jboolean JNICALL Java_io_questdb_std_Files_setLastModified
-        (JNIEnv *e, jclass cl, jlong lpszName, jlong millis) {
-    struct timeval t[2];
-    gettimeofday(t, NULL);
-    t[1].tv_sec = millis / 1000;
-    t[1].tv_usec = ((millis % 1000) * 1000);
-    return (jboolean) (utimes((const char *) lpszName, t) == 0);
-}
-
 #endif
 
 JNIEXPORT jlong JNICALL Java_io_questdb_std_Files_getStdOutFd
@@ -304,32 +288,6 @@ JNIEXPORT jboolean JNICALL Java_io_questdb_std_Files_allocate
         }
     }
     return ftruncate((int) fd, len) == 0;
-}
-
-#else
-
-JNIEXPORT jboolean JNICALL Java_io_questdb_std_Files_allocate
-        (JNIEnv *e, jclass cl, jlong fd, jlong len) {
-    int rc = posix_fallocate(fd, 0, len);
-    if (rc == 0) {
-        return JNI_TRUE;
-    }
-    if (rc == EINVAL) {
-        // Some file systems (such as ZFS) do not support posix_fallocate
-        struct stat st;
-        rc = fstat((int) fd, &st);
-        if (rc != 0) {
-            return JNI_FALSE;
-        }
-        if (st.st_size < len) {
-            rc = ftruncate(fd, len);
-            if (rc != 0) {
-                return JNI_FALSE;
-            }
-        }
-        return JNI_TRUE;
-    }
-    return JNI_FALSE;
 }
 
 #endif
@@ -502,6 +460,8 @@ JNIEXPORT jboolean JNICALL Java_io_questdb_std_Files_exists0
     return access((const char *) lpsz, F_OK) == 0;
 }
 
+#ifdef __APPLE__
+
 JNIEXPORT jlong JNICALL Java_io_questdb_std_Files_getDiskSize(JNIEnv *e, jclass cl, jlong lpszPath) {
     struct statfs buf;
     if (statfs((const char *) lpszPath, &buf) == 0) {
@@ -510,3 +470,5 @@ JNIEXPORT jlong JNICALL Java_io_questdb_std_Files_getDiskSize(JNIEnv *e, jclass 
     }
     return -1;
 }
+
+#endif
